Moved stories to their component's folder
