+++ src/Avatar/Avatar.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Story, Meta } from '@storybook/react';
-import Avatar , { AvatarProps } from '../src/Avatar';
-import { textControl, colorControl, noControl, objectControl } from './shared/utils';
+import Avatar , { AvatarProps } from '.';
+import { textControl, colorControl, noControl, objectControl } from '../shared/utils';
 
 export default {
   title: "@e-toast/Avatar",
